Split src/lib/db/attendees.ts into focused submodules
Moves the 888-line attendees DB file into separate files under
src/lib/db/attendees/ grouped by context (pii, queries, stats, capacity,
create, delete, update). The original path becomes a thin re-export shell
so existing callers keep working.
